Add Chapter 2
Chapter 2 involves the creation of a test image in the PPM format.

The PPM file format consists of a header in the following format:

P3
IMAGE_WIDTH IMAGE_HEIGHT
MAXIMUM_BRIGHTNESS_VALUE

From then on the image takes a pixel in from left to right, top to
bottom in the following order, with each pixel being on a new line:

R G B

The image render progress is printed to the stderr stream.