Add Chapter 3.
Chapter 3 involves the creation of the Vector class.

This class is used to represent both positions and colours and is the
core of the raytracing program.

Adds operations to allow for addition, subtraction,
division, multiplication and accessing via index.

In addition, unit tests are added to test the functionality of each of
the vector methods and can be executed via cargo test.