Add Chapter four.
Chapter four involves the creation of the ray class, another core
component of a raytracer (hence the name).

The ray class contains only one function at present, which calculates
a position based on an origin and direction defined upon creation,
and a t parameter to choose which point to return along the ray.

Modifies the Vec3 class to allow for bidirectional add,
subtract and multiply operations.

Chapter four uses these rays to render a background using a virtual
camera and lerping between two different colours based on ray direction.